feat: move content to /blog
