+++ src/components/layout.js
@@ -23,7 +23,7 @@ const Layout = ({ location, title, children }) => {
             textDecoration: `none`,
             color: `inherit`,
           }}
-          to={`/`}
+          to={`/blog`}
         >
           {title}
         </Link>
@@ -42,7 +42,7 @@ const Layout = ({ location, title, children }) => {
             textDecoration: `none`,
             color: `inherit`,
           }}
-          to={`/`}
+          to={`/blog`}
         >
           {title}
         </Link>
